👍Refactor 共有ボタンをコンポーネント化
