+++ src/app/_shared/template-parts/share-buttons/share-buttons.component.html
@@ -1,0 +1,6 @@
+<ul class="social-share">
+  <li><a href="{{shareURL?.twitter}}" title="ツイート" target="_blank" rel="noopener"><button ion-button block><ion-icon name="logo-twitter"></ion-icon> ツイート</button></a></li>
+  <li><button ion-button (click)="addClipboard();" block><ion-icon name="clipboard"></ion-icon> URLコピー</button></li>
+</ul>
+
+
