+++ src/app/single/single.page.html
@@ -60,8 +60,9 @@
 </ion-content>
 
 <ion-footer>
-  <ul class="social-share">
+  <!-- <ul class="social-share">
     <li><a href="{{shareURL?.twitter}}" title="ツイート" target="_blank" rel="noopener"><button ion-button block><ion-icon name="logo-twitter"></ion-icon> ツイート</button></a></li>
     <li><button ion-button (click)="addClipboard();" block><ion-icon name="clipboard"></ion-icon> URLコピー</button></li>
-  </ul>
+  </ul> -->
+  <app-share-buttons [pastData]='pastData'></app-share-buttons>
 </ion-footer>
