+++ src/app/single/single.page.scss
@@ -116,33 +116,3 @@ article.article-content {
     }
   }
 }
-ul.social-share {
-  list-style-type:none;
-  display:flex;
-  margin:0;padding:0;
-  li {
-    width:50%;
-    border-right: 1px solid #FFF;
-    background: var(--ion-color-primary);
-    button,a{
-      border-radius:0;
-      font-size: .8rem;
-      margin:0;
-      min-height:42px;
-      color: #FFF;
-      background: none;
-      width: 100%;
-      display:flex;
-      justify-content: center;
-      line-height: 1;
-      text-decoration:none;
-      ion-icon{
-        position:relative;
-        margin-right:0.5rem;
-      }
-    }
-    &:last-child{
-      border-right: none;
-    }
-  }
-}
